Refactor tests and logic
Create 'Location' class to simplify working with coordinates.
Using strings to store coordinates is good, but extracting it to
another class improves readability.

Change the way of storing ship locations.
Two separate lists for storing ship locations and hit locations are
unnecessary duplication of data.
A better way is to create a Dictionary with location as a key
and condition as value.

Adapt 'Game' class to use 'Location' class.

Implement console support in 'Program' class.

Add tests for each newly created method.